toxic_docs_insight/requirements.txt
appnope==0.1.0
boto==2.39.0
bz2file==0.98
CacheControl==0.11.6
contextlib2==0.5.1
cycler==0.10.0
decorator==4.0.9
Flask==0.10.1
funcy==1.6
future==0.15.2
gensim==0.12.4
gnureadline==6.3.3
httpretty==0.8.10
ipython==4.1.1
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
joblib==0.9.4
lockfile==0.12.2
MarkupSafe==0.23
matplotlib==1.5.1
meld3==1.0.2
natsort==4.0.4
nltk==3.1
nose==1.3.7
numexpr==2.5
numpy==1.10.4
pandas==0.17.1
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
ptyprocess==0.5.1
py==1.4.31
pyenchant==1.6.6
pyLDAvis==1.4.1
pymongo==3.2.1
pyparsing==2.1.0
pytest==2.8.7
python-dateutil==2.4.2
pytz==2015.7
requests==2.8.1
scikit-bio==0.4.1
scipy==0.17.0
seaborn==0.7.0
simplegeneric==0.8.1
six==1.10.0
smart-open==1.3.2
supervisor==3.2.1
traitlets==4.1.0
Werkzeug==0.11.4
wheel==0.29.0
